refactor(tests): consolidate unit tests and add shared fixtures
- Extract shared test fixtures (gateway-data.ts, mock-client.ts) to eliminate
  duplicate mock setup across 8+ test files
- Merge domain entity + service test pairs into 3 files: notes, posts, profiles
- Fix substack-client.test.ts to use MockService.mock.instances[0] instead of
  brittle (client as any) reassignment pattern
- Remove decodeEither from validation.ts (unused); update io-ts-validation tests
- Delete entities.test.ts (redundant with individual entity tests)
- Fix e2e error tests to use rejects.toThrow() instead of try/catch
- Update eslint.config.js to include tests/unit/fixtures/ in TypeScript parsing
