+++ tests/unit/fixtures/index.ts
@@ -1,0 +1,8 @@
+export { createMockHttpClient } from '@test/unit/fixtures/mock-client'
+export {
+  makeGatewayProfile,
+  makeGatewayNote,
+  makeGatewayPost,
+  makeGatewayFullPost,
+  makeGatewayComment
+} from '@test/unit/fixtures/gateway-data'
