+++ tests/unit/posts.test.ts
@@ -1,36 +1,20 @@
 import { FullPost, PreviewPost } from '@substack-api/domain/post'
 import { Comment } from '@substack-api/domain/comment'
+import { HttpClient } from '@substack-api/internal/http-client'
 import { CommentService } from '@substack-api/internal/services/comment-service'
 import { PostService } from '@substack-api/internal/services/post-service'
+import {
+  createMockHttpClient,
+  makeGatewayPost,
+  makeGatewayFullPost,
+  makeGatewayComment
+} from '@test/unit/fixtures'
 
-const makeGatewayPost = (id: number, title: string) => ({
-  id,
-  title,
-  subtitle: 'Test subtitle',
-  truncated_body: 'Truncated...',
-  published_at: '2023-01-01T00:00:00Z'
-})
+jest.mock('@substack-api/internal/http-client')
 
-const makeGatewayFullPost = (id: number, title: string) => ({
-  id,
-  title,
-  slug: `post-${id}`,
-  subtitle: 'Test subtitle',
-  url: `https://example.com/post-${id}`,
-  published_at: '2023-01-01T00:00:00Z',
-  html_body: '<p>Full HTML content with <strong>formatting</strong></p>',
-  truncated_body: 'Truncated content',
-  reactions: { '❤️': 5 },
-  restacks: 2,
-  tags: ['tag1'],
-  cover_image: 'https://example.com/cover.jpg'
-})
-
-const makeGatewayComment = (id: number, body: string) => ({
-  id,
-  body,
-  is_admin: false
-})
+// ---------------------------------------------------------------------------
+// PreviewPost entity
+// ---------------------------------------------------------------------------
 
 describe('PreviewPost Entity', () => {
   let mockCommentService: jest.Mocked<CommentService>
@@ -38,10 +22,7 @@ describe('PreviewPost Entity', () => {
   let post: PreviewPost
 
   beforeEach(() => {
-    mockCommentService = {
-      getCommentsForPost: jest.fn()
-    } as unknown as jest.Mocked<CommentService>
-
+    mockCommentService = { getCommentsForPost: jest.fn() } as unknown as jest.Mocked<CommentService>
     mockPostService = {
       getPostById: jest.fn(),
       getPostsForProfile: jest.fn()
@@ -51,21 +32,15 @@ describe('PreviewPost Entity', () => {
   })
 
   describe('properties', () => {
-    it('should expose id', () => {
+    it('should expose id, title, and publishedAt', () => {
       expect(post.id).toBe(456)
-    })
-
-    it('should expose title', () => {
       expect(post.title).toBe('Test Post')
-    })
-
-    it('should expose publishedAt as Date', () => {
       expect(post.publishedAt).toBeInstanceOf(Date)
     })
   })
 
   describe('comments()', () => {
-    it('should iterate through post comments', async () => {
+    it('should iterate through post comments as Comment instances', async () => {
       const mockComments = [makeGatewayComment(1, 'Comment 1'), makeGatewayComment(2, 'Comment 2')]
       mockCommentService.getCommentsForPost.mockResolvedValue(mockComments)
 
@@ -93,7 +68,6 @@ describe('PreviewPost Entity', () => {
       }
 
       expect(comments).toHaveLength(1)
-      expect(comments[0].body).toBe('Comment 1')
     })
 
     it('should handle empty comments', async () => {
@@ -127,7 +101,6 @@ describe('PreviewPost Entity', () => {
 
       expect(fullPost).toBeInstanceOf(FullPost)
       expect(fullPost.id).toBe(456)
-      expect(fullPost.title).toBe('Test Post')
       expect(mockPostService.getPostById).toHaveBeenCalledWith(456)
     })
 
@@ -137,58 +110,97 @@ describe('PreviewPost Entity', () => {
     })
   })
 
-  describe('like()', () => {
+  describe('like() and addComment()', () => {
     it('should throw not-implemented error', async () => {
       await expect(post.like()).rejects.toThrow()
-    })
-  })
-
-  describe('addComment()', () => {
-    it('should throw not-implemented error', async () => {
       await expect(post.addComment({ body: 'Test' })).rejects.toThrow()
     })
   })
 })
+
+// ---------------------------------------------------------------------------
+// FullPost entity
+// ---------------------------------------------------------------------------
 
 describe('FullPost Entity', () => {
   let mockCommentService: jest.Mocked<CommentService>
   let fullPost: FullPost
 
   beforeEach(() => {
-    mockCommentService = {
-      getCommentsForPost: jest.fn()
-    } as unknown as jest.Mocked<CommentService>
-
+    mockCommentService = { getCommentsForPost: jest.fn() } as unknown as jest.Mocked<CommentService>
     fullPost = new FullPost(makeGatewayFullPost(789, 'Full Test Post'), mockCommentService)
   })
 
   describe('properties', () => {
-    it('should expose id', () => {
+    it('should expose id, title, htmlBody, subtitle, and publishedAt', () => {
       expect(fullPost.id).toBe(789)
-    })
-
-    it('should expose title', () => {
       expect(fullPost.title).toBe('Full Test Post')
-    })
-
-    it('should expose htmlBody', () => {
       expect(fullPost.htmlBody).toBe('<p>Full HTML content with <strong>formatting</strong></p>')
-    })
-
-    it('should expose subtitle', () => {
       expect(fullPost.subtitle).toBe('Test subtitle')
-    })
-
-    it('should expose publishedAt as Date', () => {
       expect(fullPost.publishedAt).toBeInstanceOf(Date)
     })
-  })
 
-  describe('Post interface', () => {
     it('should have like, addComment, and comments methods', () => {
       expect(typeof fullPost.like).toBe('function')
       expect(typeof fullPost.addComment).toBe('function')
       expect(typeof fullPost.comments).toBe('function')
+    })
+  })
+})
+
+// ---------------------------------------------------------------------------
+// PostService
+// ---------------------------------------------------------------------------
+
+describe('PostService', () => {
+  let postService: PostService
+  let mockClient: jest.Mocked<HttpClient>
+
+  beforeEach(() => {
+    jest.clearAllMocks()
+    mockClient = createMockHttpClient()
+    postService = new PostService(mockClient)
+  })
+
+  describe('getPostById', () => {
+    it('should return GatewayFullPost from GET /posts/{id}', async () => {
+      const mockPost = makeGatewayFullPost(123, 'Test Post')
+      mockClient.get.mockResolvedValueOnce(mockPost)
+
+      expect(await postService.getPostById(123)).toEqual(mockPost)
+      expect(mockClient.get).toHaveBeenCalledWith('/posts/123')
+    })
+
+    it('should throw when request fails', async () => {
+      mockClient.get.mockRejectedValueOnce(new Error('HTTP 404: Not found'))
+      await expect(postService.getPostById(999)).rejects.toThrow('HTTP 404: Not found')
+    })
+  })
+
+  describe('getPostsForProfile', () => {
+    it('should return posts from GET /profiles/{slug}/posts', async () => {
+      const mockPosts = [makeGatewayPost(1, 'Post 1'), makeGatewayPost(2, 'Post 2')]
+      mockClient.get.mockResolvedValueOnce({ items: mockPosts })
+
+      const result = await postService.getPostsForProfile('testuser', { limit: 10, offset: 0 })
+
+      expect(result).toEqual(mockPosts)
+      expect(mockClient.get).toHaveBeenCalledWith('/profiles/testuser/posts', {
+        limit: 10,
+        offset: 0
+      })
+    })
+
+    it('should return empty array when profile has no posts', async () => {
+      mockClient.get.mockResolvedValueOnce({ items: [] })
+      expect(await postService.getPostsForProfile('testuser', { limit: 5, offset: 0 })).toEqual([])
+    })
+
+    it('should throw when request fails', async () => {
+      mockClient.get.mockRejectedValueOnce(new Error('HTTP 500: Internal server error'))
+      await expect(
+        postService.getPostsForProfile('testuser', { limit: 10, offset: 0 })
+      ).rejects.toThrow('HTTP 500: Internal server error')
     })
   })
 })
