+++ tests/unit/substack-client.test.ts
@@ -5,22 +5,23 @@ import {
   PostService,
   NoteService,
   ProfileService,
-  CommentService,
-  FollowingService,
   ConnectivityService
 } from '@substack-api/internal/services'
+import { makeGatewayNote } from '@test/unit/fixtures'
 
 jest.mock('@substack-api/internal/http-client')
 jest.mock('@substack-api/internal/services')
 
+const MockPostService = PostService as jest.MockedClass<typeof PostService>
+const MockNoteService = NoteService as jest.MockedClass<typeof NoteService>
+const MockProfileService = ProfileService as jest.MockedClass<typeof ProfileService>
+const MockConnectivityService = ConnectivityService as jest.MockedClass<typeof ConnectivityService>
+
 describe('SubstackClient', () => {
   let client: SubstackClient
-  let mockClient: jest.Mocked<HttpClient>
   let mockPostService: jest.Mocked<PostService>
   let mockNoteService: jest.Mocked<NoteService>
   let mockProfileService: jest.Mocked<ProfileService>
-  let mockCommentService: jest.Mocked<CommentService>
-  let mockFollowingService: jest.Mocked<FollowingService>
   let mockConnectivityService: jest.Mocked<ConnectivityService>
 
   const gatewayConfig = {
@@ -29,87 +30,50 @@ describe('SubstackClient', () => {
     token: 'dummy-token'
   }
 
-  const makeGatewayNote = (id: number, body: string) => ({
-    id,
-    body,
-    likes_count: 0,
-    author: { id: 123, name: 'Test User', handle: 'testuser', avatar_url: '' },
-    published_at: '2023-01-01T00:00:00Z'
-  })
-
   beforeEach(() => {
     jest.clearAllMocks()
-
-    mockClient = new HttpClient('https://test.com', {
-      token: 'dummy-token',
-      publicationUrl: 'https://pub.com'
-    }) as jest.Mocked<HttpClient>
-    mockClient.get = jest.fn()
-    mockClient.post = jest.fn()
-
-    mockPostService = new PostService(mockClient) as jest.Mocked<PostService>
-    mockPostService.getPostById = jest.fn()
-
-    mockNoteService = new NoteService(mockClient) as jest.Mocked<NoteService>
-    mockNoteService.getNoteById = jest.fn()
-
-    mockProfileService = new ProfileService(mockClient) as jest.Mocked<ProfileService>
-    mockProfileService.getOwnProfile = jest.fn()
-    mockProfileService.getProfileBySlug = jest.fn()
-
-    mockCommentService = new CommentService(mockClient) as jest.Mocked<CommentService>
-    mockCommentService.getCommentsForPost = jest.fn()
-
-    mockFollowingService = new FollowingService(mockClient) as jest.Mocked<FollowingService>
-    mockFollowingService.getFollowing = jest.fn()
-
-    mockConnectivityService = new ConnectivityService(
-      mockClient
-    ) as jest.Mocked<ConnectivityService>
-    mockConnectivityService.isConnected = jest.fn()
-
     client = new SubstackClient(gatewayConfig)
-    ;(client as any).postService = mockPostService
-    ;(client as any).noteService = mockNoteService
-    ;(client as any).profileService = mockProfileService
-    ;(client as any).commentService = mockCommentService
-    ;(client as any).followingService = mockFollowingService
-    ;(client as any).connectivityService = mockConnectivityService
+
+    // Retrieve service instances created inside the SubstackClient constructor
+    mockPostService = MockPostService.mock.instances[0] as jest.Mocked<PostService>
+    mockNoteService = MockNoteService.mock.instances[0] as jest.Mocked<NoteService>
+    mockProfileService = MockProfileService.mock.instances[0] as jest.Mocked<ProfileService>
+    mockConnectivityService = MockConnectivityService.mock
+      .instances[0] as jest.Mocked<ConnectivityService>
   })
 
   describe('constructor', () => {
-    it('should create client instance with gateway config', () => {
+    it('should create client instance', () => {
       expect(client).toBeInstanceOf(SubstackClient)
     })
 
-    it('should construct HttpClient with gatewayUrl', () => {
+    it('should construct HttpClient with the provided gatewayUrl', () => {
       jest.clearAllMocks()
       new SubstackClient(gatewayConfig)
-      const calls = (HttpClient as jest.MockedClass<typeof HttpClient>).mock.calls
-      expect(calls[0][0]).toContain('localhost:5001')
+      expect((HttpClient as jest.MockedClass<typeof HttpClient>).mock.calls[0][0]).toContain(
+        'localhost:5001'
+      )
     })
 
     it('should use default gateway URL when gatewayUrl is not provided', () => {
       jest.clearAllMocks()
       const { gatewayUrl: _, ...configWithoutUrl } = gatewayConfig
       new SubstackClient(configWithoutUrl)
-      const calls = (HttpClient as jest.MockedClass<typeof HttpClient>).mock.calls
-      expect(calls[0][0]).toContain('substack-gateway.vercel.app')
+      expect((HttpClient as jest.MockedClass<typeof HttpClient>).mock.calls[0][0]).toContain(
+        'substack-gateway.vercel.app'
+      )
     })
   })
 
   describe('testConnectivity', () => {
     it('should return true when gateway is accessible', async () => {
       mockConnectivityService.isConnected.mockResolvedValue(true)
-      const result = await client.testConnectivity()
-      expect(result).toBe(true)
-      expect(mockConnectivityService.isConnected).toHaveBeenCalled()
+      expect(await client.testConnectivity()).toBe(true)
     })
 
     it('should return false when gateway is not accessible', async () => {
       mockConnectivityService.isConnected.mockResolvedValue(false)
-      const result = await client.testConnectivity()
-      expect(result).toBe(false)
+      expect(await client.testConnectivity()).toBe(false)
     })
   })
 
@@ -127,11 +91,9 @@ describe('SubstackClient', () => {
       const result = await client.ownProfile()
       expect(result).toBeInstanceOf(OwnProfile)
       expect(result.id).toBe(123)
-      expect(result.name).toBe('Test User')
-      expect(mockProfileService.getOwnProfile).toHaveBeenCalled()
     })
 
-    it('should throw error when authentication fails', async () => {
+    it('should throw when authentication fails', async () => {
       mockProfileService.getOwnProfile.mockRejectedValue(new Error('Unauthorized'))
       await expect(client.ownProfile()).rejects.toThrow('Failed to get own profile: Unauthorized')
     })
@@ -161,8 +123,7 @@ describe('SubstackClient', () => {
     it('should throw when profile not found', async () => {
       mockProfileService.getProfileBySlug.mockRejectedValue(new Error('Not found'))
       await expect(client.profileForSlug('nonexistent')).rejects.toThrow(
-        // eslint-disable-next-line quotes
-        "Profile with slug 'nonexistent' not found: Not found"
+        /Profile with slug.*nonexistent.*not found/
       )
     })
   })
@@ -199,7 +160,6 @@ describe('SubstackClient', () => {
       expect(note).toBeInstanceOf(Note)
       expect(note.id).toBe(789)
       expect(note.body).toBe('Test note')
-      expect(mockNoteService.getNoteById).toHaveBeenCalledWith(789)
     })
 
     it('should throw when note not found', async () => {
